Module Basic.Ack

When sent by the client, this method acknowledges one or more messages delivered via the Deliver or Get-Ok methods.

When sent by server, this method acknowledges one or more messages published with the Publish method on a channel in confirm mode.

The acknowledgement can be for a single message or a set of messages up to and including a specific message.

type t = {

delivery_tag : delivery_tag;

multiple : Amqp_client_lib.Types.bit;
(*
If set to 1, the delivery tag is treated as "up to and including", so that multiple messages can be acknowledged with a single method. If set to zero, the delivery tag refers to a single message. If the multiple field is 1, and the delivery tag is zero, this indicates acknowledgement of all outstanding messages.
*)

}

val init : delivery_tag:delivery_tag -> multiple:Amqp_client_lib.Types.bit -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> t Lwt.t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t