Module Basic.Consume

This method asks the server to start a "consumer", which is a transient request for messages from a specific queue. Consumers last as long as the channel they were declared on, or until the client cancels them.

type t = {

queue : queue_name;
(*
Specifies the name of the queue to consume from.
*)

consumer_tag : consumer_tag;
(*
Specifies the identifier for the consumer. The consumer tag is local to a channel, so two clients can use the same consumer tags. If this field is empty the server will generate a unique tag.
*)

no_local : no_local;

no_ack : no_ack;

exclusive : Amqp_client_lib.Types.bit;
(*
Request exclusive consumer access, meaning only this consumer can access the queue.
*)

no_wait : no_wait;

arguments : Amqp_client_lib.Types.table;
(*
A set of arguments for the consume. The syntax and semantics of these arguments depends on the server implementation.
*)

}

val init : queue:'a -> consumer_tag:queue_name -> no_local:consumer_tag -> no_ack:no_local -> exclusive:no_ack -> no_wait:Amqp_client_lib.Types.bit -> arguments:no_wait -> unit -> Amqp_client_lib.Types.table -> t

val request : (Framing.t * Framing.channel_no) -> t -> Consume_ok.t Thread.Deferred.t