Module Basic.Get_ok

This method delivers a message to the client following a get method. A message delivered by 'get-ok' must be acknowledged unless the no-ack option was set in the get method.

type t = {

delivery_tag : delivery_tag;

redelivered : redelivered;

exchange : exchange_name;
(*
Specifies the name of the exchange that the message was originally published to. If empty, the message was published to the default exchange.
*)

routing_key : Amqp_client_lib.Types.shortstr;
(*
Specifies the routing key name specified when the message was published.
*)

message_count : message_count;

}

val init : delivery_tag:delivery_tag -> redelivered:redelivered -> exchange:exchange_name -> routing_key:Amqp_client_lib.Types.shortstr -> message_count:message_count -> unit -> t