Module Basic.Return

This method returns an undeliverable message that was published with the "immediate" flag set, or an unroutable message published with the "mandatory" flag set. The reply code and text provide information about the reason that the message was undeliverable.

type t = {

reply_code : reply_code;

reply_text : reply_text;

exchange : exchange_name;
(*
Specifies the name of the exchange that the message was originally published to. May be empty, meaning the default exchange.
*)

routing_key : Amqp_client_lib.Types.shortstr;
(*
Specifies the routing key name specified when the message was published.
*)

}

val init : reply_code:reply_code -> reply_text:reply_text -> exchange:exchange_name -> routing_key:Amqp_client_lib.Types.shortstr -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t * (Content.t * string)) Lwt.t