Module Spec.Connection

The connection class provides methods for a client to establish a network connection to a server, and for both peers to operate the connection thereafter.

module Start_ok : sig ... end

This method selects a SASL security mechanism.

module Start : sig ... end

This method starts the connection negotiation process by telling the client the protocol version that the server proposes, along with a list of security mechanisms which the client can use for authentication.

module Secure_ok : sig ... end

This method attempts to authenticate, passing a block of SASL data for the security mechanism at the server side.

module Secure : sig ... end

The SASL protocol works by exchanging challenges and responses until both peers have received sufficient information to authenticate each other. This method challenges the client to provide more information.

module Tune_ok : sig ... end

This method sends the client's connection tuning parameters to the server. Certain fields are negotiated, others provide capability information.

module Tune : sig ... end

This method proposes a set of connection configuration values to the client. The client can accept and/or adjust these.

module Open_ok : sig ... end

This method signals to the client that the connection is ready for use.

module Open : sig ... end

This method opens a connection to a virtual host, which is a collection of resources, and acts to separate multiple application domains within a server. The server may apply arbitrary limits per virtual host, such as the number of each type of entity that may be used, per connection and/or in total.

module Close_ok : sig ... end

This method confirms a Connection.Close method and tells the recipient that it is safe to release resources for the connection and close the socket.

module Close : sig ... end

This method indicates that the sender wants to close the connection. This may be due to internal conditions (e.g. a forced shut-down) or due to an error handling a specific method, i.e. an exception. When a close is due to an exception, the sender provides the class and method id of the method which caused the exception.

module Blocked : sig ... end

This method indicates that a connection has been blocked and does not accept new publishes.

module Unblocked : sig ... end

This method indicates that a connection has been unblocked and now accepts publishes.

module Update_secret_ok : sig ... end

This method confirms the updated secret is valid.

module Update_secret : sig ... end

This method updates the secret used to authenticate this connection. It is used when secrets have an expiration date and need to be renewed, like OAuth 2 tokens.