Module Connection.Close

This method indicates that the sender wants to close the connection. This may be due to internal conditions (e.g. a forced shut-down) or due to an error handling a specific method, i.e. an exception. When a close is due to an exception, the sender provides the class and method id of the method which caused the exception.

type t = {

reply_code : reply_code;

reply_text : reply_text;

class_id : class_id;
(*
When the close is provoked by a method exception, this is the class of the method.
*)

method_id : method_id;
(*
When the close is provoked by a method exception, this is the ID of the method.
*)

}

val init : reply_code:reply_code -> reply_text:reply_text -> class_id:class_id -> method_id:method_id -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t -> unit Thread.Deferred.t) -> unit Thread.Deferred.t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t