Module Connection.Close_ok

This method confirms a Connection.Close method and tells the recipient that it is safe to release resources for the connection and close the socket.

type t = unit

val init : unit -> unit