Module Connection.Open

This method opens a connection to a virtual host, which is a collection of resources, and acts to separate multiple application domains within a server. The server may apply arbitrary limits per virtual host, such as the number of each type of entity that may be used, per connection and/or in total.

type t = {

virtual_host : path;
(*
The name of the virtual host to work with.
*)

}

val init : virtual_host:path -> unit -> 'a -> 'b -> t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t