Module Connection.Secure

The SASL protocol works by exchanging challenges and responses until both peers have received sufficient information to authenticate each other. This method challenges the client to provide more information.

type t = {

challenge : Amqp_client_lib.Types.longstr;
(*
Challenge information, a block of opaque binary data passed to the security mechanism.
*)

}

val init : challenge:Amqp_client_lib.Types.longstr -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t -> Secure_ok.t Thread.Deferred.t) -> unit Thread.Deferred.t