Module Connection.Tune

This method proposes a set of connection configuration values to the client. The client can accept and/or adjust these.

type t = {

channel_max : Amqp_client_lib.Types.short;
(*
Specifies highest channel number that the server permits. Usable channel numbers are in the range 1..channel-max. Zero indicates no specified limit.
*)

frame_max : Amqp_client_lib.Types.long;
(*
The largest frame size that the server proposes for the connection, including frame header and end-byte. The client can negotiate a lower value. Zero means that the server does not impose any specific limit but may reject very large frames if it cannot allocate resources for them.
*)

heartbeat : Amqp_client_lib.Types.short;
(*
The delay, in seconds, of the connection heartbeat that the server wants. Zero means the server does not want a heartbeat.
*)

}

val init : channel_max:Amqp_client_lib.Types.short -> frame_max:Amqp_client_lib.Types.long -> heartbeat:Amqp_client_lib.Types.short -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t -> Tune_ok.t Thread.Deferred.t) -> unit Thread.Deferred.t