Module Connection.Update_secret

This method updates the secret used to authenticate this connection. It is used when secrets have an expiration date and need to be renewed, like OAuth 2 tokens.

type t = {

new_secret : Amqp_client_lib.Types.longstr;
(*
The new secret.
*)

reason : Amqp_client_lib.Types.shortstr;
(*
The reason for the secret update.
*)

}

val init : new_secret:Amqp_client_lib.Types.longstr -> reason:Amqp_client_lib.Types.shortstr -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t -> unit Thread.Deferred.t) -> unit Thread.Deferred.t