Module Spec.Queue

Queues store and forward messages. Queues can be configured in the server or created at runtime. Queues must be attached to at least one exchange in order to receive messages from publishers.

module Declare_ok : sig ... end

This method confirms a Declare method and confirms the name of the queue, essential for automatically-named queues.

module Declare : sig ... end

This method creates or checks a queue. When creating a new queue the client can specify various properties that control the durability of the queue and its contents, and the level of sharing for the queue.

module Bind_ok : sig ... end

This method confirms that the bind was successful.

module Bind : sig ... end

This method binds a queue to an exchange. Until a queue is bound it will not receive any messages. In a classic messaging model, store-and-forward queues are bound to a direct exchange and subscription queues are bound to a topic exchange.

module Unbind_ok : sig ... end

This method confirms that the unbind was successful.

module Unbind : sig ... end

This method unbinds a queue from an exchange.

module Purge_ok : sig ... end

This method confirms the purge of a queue.

module Purge : sig ... end

This method removes all messages from a queue which are not awaiting acknowledgment.

module Delete_ok : sig ... end

This method confirms the deletion of a queue.

module Delete : sig ... end

This method deletes a queue. When a queue is deleted any pending messages are sent to a dead-letter queue if this is defined in the server configuration, and all consumers on the queue are cancelled.