Module Channel.Flow

This method asks the peer to pause or restart the flow of content data sent by a consumer. This is a simple flow-control mechanism that a peer can use to avoid overflowing its queues or otherwise finding itself receiving more messages than it can process. Note that this method is not intended for window control. It does not affect contents returned by Basic.Get-Ok methods.

type t = {

active : Amqp_client_lib.Types.bit;
If 1, the peer starts sending content frames. If 0, the peer stops sending content frames.

}

val init : active:Amqp_client_lib.Types.bit -> unit -> t

val reply : ?⁠once:bool -> (Amqp_client_lwt.Framing.t * Amqp_client_lwt.Framing.channel_no) -> (t -> Flow_ok.t Amqp_client_lwt.Thread.Deferred.t) -> unit Amqp_client_lwt.Thread.Deferred.t

val request : (Amqp_client_lwt.Framing.t * Amqp_client_lwt.Framing.channel_no) -> t -> Flow_ok.t Amqp_client_lwt.Thread.Deferred.t