Module Exchange.Bind

This method binds an exchange to an exchange.

type t = {

destination : exchange_name;
Specifies the name of the destination exchange to bind.

source : exchange_name;
Specifies the name of the source exchange to bind.

routing_key : Amqp_client_lib.Types.shortstr;
Specifies the routing key for the binding. The routing key is used for routing messages depending on the exchange configuration. Not all exchanges use a routing key - refer to the specific exchange documentation.

no_wait : no_wait;

arguments : Amqp_client_lib.Types.table;
A set of arguments for the binding. The syntax and semantics of these arguments depends on the exchange class.

}

val init : destination:'a -> source:exchange_name -> routing_key:exchange_name -> no_wait:Amqp_client_lib.Types.shortstr -> arguments:no_wait -> unit -> Amqp_client_lib.Types.table -> t

val request : (Amqp_client_lwt.Framing.t * Amqp_client_lwt.Framing.channel_no) -> t -> unit Amqp_client_lwt.Thread.Deferred.t